pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visions"
dynamic = ["version"]
description = "Visions"
readme = "README.md"
license = { text = "BSD License" }
authors = [
    { name = "Dylan Profiler", email = "visions@ictopzee.nl" }
]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "numpy>=1.23.2",
    "pandas>=2.0.0",
    "attrs>=19.3.0",
    "networkx>=2.4",
    "multimethod>=1.4",
    "puremagic",
]

[project.optional-dependencies]
type_geometry = [
    "shapely"
]
type_image_path = [
    "imagehash",
    "Pillow",
    "puremagic"
]
plotting = [
    "pydot",
    "pygraphviz",
    "matplotlib"
]
testing = [
    "visions[type_geometry, type_image_path]",
    "setuptools>=46.1.3",
    "wheel>=0.34.2",
    "black>=20.8b1",
    "isort>=5.0.9",
    "mypy>=0.770",
    "recommonmark>=0.6.0",
    "nbsphinx",
    "sphinx_rtd_theme>=0.4.3",
    "sphinx-autodoc-typehints>=1.10.3",
    "IPython",
    "Sphinx-copybutton",
    "big_o>=0.10.1",
    "twine>=3.1.1",
    "pydot",
    "pytest>=5.2.0",
    "check-manifest>=0.41",
    "pre-commit",
    "pytest-spark>=0.6.0",
    "pyarrow>=1.0.1",
    "numba",
    "pyspark",
]

dev = [
    "visions[testing, plotting]"
]

all = [
    "visions[type_geometry, type_image_path, plotting]"
]


[project.urls]
Repository = "https://github.com/dylan-profiler/visions"
Documentation = "https://dylan-profiler.github.io/visions"


[tool.setuptools]
packages = ["visions"]
package-data = { vision = ["py.typed"] }
include-package-data = true
zip-safe = false

[tool.setuptools.package-dir]
visions = "src/visions"

[tool.setuptools.dynamic]
version = { attr = "visions.__version__" }